Refactor environment constants: rename file and update import paths for consistency; remove old env-constants.ts references.
